Add deposit validation, abstraction over console input and constant stings and other thinks

+++ BedeGaming.Simple.Slot.Machine/BedeGaming.SimpleSlotMachine.Application/Interfaces/ISlotMachineService.cs
@@ -2,6 +2,6 @@
 {
     public interface ISlotMachineService
     {
-        void Play(int stakeAmount);
+        void Play(double stakeAmount);
     }
 }
